Remove redundant int conversion
In Python3 `round`, `math.floor`, and `math.ceil` always return an
integer, so it is no more necessary to convert the result explicitly to
integer.

+++ lib/matplotlib/image.py
@@ -1190,8 +1190,8 @@ class PcolorImage(AxesImage):
         l, b, r, t = self.axes.bbox.extents
         width = (round(r) + 0.5) - (round(l) - 0.5)
         height = (round(t) + 0.5) - (round(b) - 0.5)
-        width = int(round(width * magnification))
-        height = int(round(height * magnification))
+        width = round(width * magnification)
+        height = round(height * magnification)
         if self._rgbacache is None:
             A = self.to_rgba(self._A, bytes=True)
             self._rgbacache = A
